feat(vertex_ai): Support task type and title in embeddings models
The embeddings API now support a "task type" parameter to define the kind of task the embedding will be used for to help the model produce better quality embeddings.

- RETRIEVAL_QUERY: Specifies the given text is a query in a search/retrieval setting.
- RETRIEVAL_DOCUMENT: Specifies the given text is a document in a search/retrieval setting.
- SEMANTIC_SIMILARITY: Specifies the given text will be used for Semantic Textual Similarity (STS).
- CLASSIFICATION: Specifies that the embeddings will be used for classification.
- CLUSTERING: Specifies that the embeddings will be used for clustering.

For RETRIEVAL_DOCUMENT task, a "title" parameter can also be attached.

These two new parameters are only supported in models released in or after August 2023.

Docs:
https://cloud.google.com/vertex-ai/docs/generative-ai/embeddings/get-text-embeddings#api_changes_to_models_released_in_or_after_august_2023