feat(embeddings): Support task type in VertexAI embeddings models
VertexAI embeddings models released in August support a new "task type" parameter that indicates the kind of task the embeddings is going to be used for, this helps the model produce better quality embeddings.

If you are using a model that supports it, LangChain.dart will use 'RETRIEVAL_DOCUMENT' when embedding a document and 'RETRIEVAL_QUERY' when embedding a query.